server/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
celery==5.4.0
channels==4.1.0
daphne==4.1.2
